Cleaning up the code a little - Tiny bugfixes to my previous PR - Change most Vector3d to Vector3f for performance - Change most stdint types (eg. uint32_t) to libctru types (eg. u32) - Shuffle around includes to prefer Forward declarations, to have the code less tangled when changing a header - ScreenFwd.h with a few common includes that can be included by every Screen TU. - Put World Conf structs/enums to WorldConf.h to avoid including the whole CT_World.h Following will be the GameSettings.

+++ source/recraft/misc/CommandLine.cpp
@@ -1,5 +1,4 @@
 #include "misc/CommandLine.hpp"
-#include "world/savegame/SaveManager.hpp"
 
 #include <3ds.h>
 
@@ -9,103 +8,104 @@
 #include <stdio.h>
 #include <string.h>
 
+#include "entity/Player.hpp"
+#include "gui/DebugUI.hpp"
+#include "world/CT_World.hpp"
+#include "world/savegame/SaveManager.hpp"
 
-//TODO: ImGUI for this (at least for now), this needs to work with online
-void CommandLine_Activate(World* world, Player* player,DebugUI* debugUi) {
-	static SwkbdState swkbd;
-	static char textBuffer[64];
-	swkbdInit(&swkbd, SWKBD_TYPE_WESTERN, 2, 64);
-	swkbdSetHintText(&swkbd, "Enter command");
+// TODO: ImGUI for this (at least for now), this needs to work with online
+void CommandLine_Activate(World* world, Player* player, DebugUI* debugUi) {
+    static SwkbdState swkbd;
+    static char       textBuffer[64];
+    swkbdInit(&swkbd, SWKBD_TYPE_WESTERN, 2, 64);
+    swkbdSetHintText(&swkbd, "Enter command");
 
-	int button = swkbdInputText(&swkbd, textBuffer, sizeof(textBuffer));
-	if (button == SWKBD_BUTTON_CONFIRM) {
-		CommandLine_Execute(world, player, debugUi, textBuffer);
-	}
+    int button = swkbdInputText(&swkbd, textBuffer, sizeof(textBuffer));
+    if (button == SWKBD_BUTTON_CONFIRM) {
+        CommandLine_Execute(world, player, debugUi, textBuffer);
+    }
 }
 
 void CommandLine_Execute(World* world, Player* player, DebugUI* debugUi, const char* text) {
-	int length = strlen(text);
-	mpack_writer_t writer;
-	mpack_error_t err = mpack_writer_destroy(&writer);
-	if (length >= 1 && text[0] == '/'&&player->cheats==true) {
-		if (length >= 9) {
-			float x, y, z;
-			if (sscanf(&text[1], "tp %f %f %f", &x, &y, &z) == 3) {
-				player->position.x = x;
-				player->position.y = y + 1;
-				player->position.z = z;
+    int            length = strlen(text);
+    mpack_writer_t writer;
+    mpack_error_t  err = mpack_writer_destroy(&writer);
+    if (length >= 1 && text[0] == '/' && player->cheats == true) {
+        if (length >= 9) {
+            float x, y, z;
+            if (sscanf(&text[1], "tp %f %f %f", &x, &y, &z) == 3) {
+                player->position.x = x;
+                player->position.y = y + 1;
+                player->position.z = z;
                 debugUi->Log("Teleported to %f, %f %f", x, y, z);
-			}
-		}
-		//int playerid;
-		if (length == 2 && text[1] == 'k') {
-			player->hp=0;
+            }
+        }
+        // int playerid;
+        if (length == 2 && text[1] == 'k') {
+            player->hp = 0;
             debugUi->Log("Killed player");
-		}
-		int hp;
-		if (sscanf(&text[1],"hp %i",&hp)) {
-			if (hp>0&&hp<21){
-				player->hp=hp;
+        }
+        int hp;
+        if (sscanf(&text[1], "hp %i", &hp)) {
+            if (hp > 0 && hp < 21) {
+                player->hp = hp;
                 debugUi->Log("Set player hp to %i", hp);
-			} else {
-                debugUi->Log("Cannot set hp to %i",hp);
-			}
-		}
-		double sx,sy,sz;
-		if (sscanf(&text[1], "ws %d %d %d", &sx, &sy, &sz) == 3) {
-			player->spawnPos.x = sx;
-			player->spawnPos.y = sy;
-			player->spawnPos.z = sz;
-			player->spawnset=true;
-			mpack_write_cstr(&writer, "sx");
-			mpack_write_double(&writer,player->spawnPos.x);
-			mpack_write_cstr(&writer, "sy");
-			mpack_write_double(&writer,player->spawnPos.y);
-			mpack_write_cstr(&writer, "sz");
-			mpack_write_double(&writer,player->spawnPos.z);
-			mpack_write_cstr(&writer, "ss");
-			mpack_write_bool(&writer,player->spawnset);
+            } else {
+                debugUi->Log("Cannot set hp to %i", hp);
+            }
+        }
+        float sx, sy, sz;
+        if (sscanf(&text[1], "ws %f %f %f", &sx, &sy, &sz) == 3) {
+            player->spawnPos.x = sx;
+            player->spawnPos.y = sy;
+            player->spawnPos.z = sz;
+            player->spawnset   = true;
+            mpack_write_cstr(&writer, "sx");
+            mpack_write_double(&writer, player->spawnPos.x);
+            mpack_write_cstr(&writer, "sy");
+            mpack_write_double(&writer, player->spawnPos.y);
+            mpack_write_cstr(&writer, "sz");
+            mpack_write_double(&writer, player->spawnPos.z);
+            mpack_write_cstr(&writer, "ss");
+            mpack_write_bool(&writer, player->spawnset);
             debugUi->Log("Set spawn to %d, %d %d", sx, sy, sz);
-			if (err != mpack_ok) {
+            if (err != mpack_ok) {
                 debugUi->Log("Mpack error %d while saving world manifest", err);
                 debugUi->Log("Save file possibly corrupted, don't hit me plz");
-			}
-		}
-		int gm;
-		if (sscanf(&text[1],"gm %i",&gm)) {
-			if (gm>0&&gm<5){
-				player->gamemode=gm;
+            }
+        }
+        int gm;
+        if (sscanf(&text[1], "gm %i", &gm)) {
+            if (gm > 0 && gm < 5) {
+                player->gamemode = gm;
                 debugUi->Log("Set gamemode to %i", gm);
-			} else {
-                debugUi->Log("Cannot set gamemode to %i",gm);
-			}
-
-		}
-		int hunger;
-		if (sscanf(&text[1],"hunger %i",&hunger)) {
-			if (hunger>0&&hunger<5){
-				player->hunger=hunger;
+            } else {
+                debugUi->Log("Cannot set gamemode to %i", gm);
+            }
+        }
+        int hunger;
+        if (sscanf(&text[1], "hunger %i", &hunger)) {
+            if (hunger > 0 && hunger < 5) {
+                player->hunger = hunger;
                 debugUi->Log("Set hunger to %i", hunger);
-			} else {
-                debugUi->Log("Cannot set hunger to %i",hunger);
-			}
-
-		}
-		int diff;
-		if (sscanf(&text[1],"diff %i",&diff)) {
-			if (diff>0&&diff<6) {
-				player->difficulty=diff;
+            } else {
+                debugUi->Log("Cannot set hunger to %i", hunger);
+            }
+        }
+        int diff;
+        if (sscanf(&text[1], "diff %i", &diff)) {
+            if (diff > 0 && diff < 6) {
+                player->difficulty = diff;
                 debugUi->Log("Set difficulty to %i", diff);
-			} else {
-                debugUi->Log("Cannot set difficulty to %i",diff);
-			}
-		}
-		/*int bx, by, bz; 					idfk how to get it to read strings as arguments, compiler isn't liking it
-		char block;
-		if (sscanf(&text[1], "sb %i %i %i", &bx, &by, &bz) == 3 &&) {
-			World_SetBlock(player->world, bx,by,bz, block);
-			DebugUI_Log("Block at %f, %f %f is now %c", bx, by, bz,block);
-		}*/
-		
-	}
+            } else {
+                debugUi->Log("Cannot set difficulty to %i", diff);
+            }
+        }
+        /*int bx, by, bz; 					idfk how to get it to read strings as arguments, compiler isn't liking it
+        char block;
+        if (sscanf(&text[1], "sb %i %i %i", &bx, &by, &bz) == 3 &&) {
+            World_SetBlock(player->world, bx,by,bz, block);
+            DebugUI_Log("Block at %f, %f %f is now %c", bx, by, bz,block);
+        }*/
+    }
 }
